Breakup typeset.ts into separate files
