+++ packages/typesetter/src/layout.ts
@@ -1,12 +1,10 @@
-import * as Editor from "@math-blocks/editor-core";
 import {UnreachableCaseError} from "@math-blocks/core";
 import type {FontData, MathValueRecord} from "@math-blocks/opentype";
 
-import {multiplierForContext, fontSizeForContext} from "./utils";
+import {fontSizeForContext} from "./utils";
 import type {Context} from "./types";
-import {MathStyle} from "./enums";
 
-type Dist = number;
+export type Dist = number;
 
 export type Dim = {
     width: Dist;
@@ -317,91 +315,7 @@ export const makeVBox = (
     return makeBox("vbox", dim, [nodeList], context);
 };
 
-const makeList = (size: Dist, box: Box): readonly Node[] => [
-    makeKern(size),
-    box,
-];
-
-// TODO: compute width from numBox and denBox
-export const makeFract = (numBox: Box, denBox: Box, context: Context): Box => {
-    const {mathStyle} = context;
-    const {constants} = context.fontData.font.math;
-
-    const fontSize = fontSizeForContext(context);
-    const thickness = (fontSize * constants.fractionRuleThickness.value) / 1000;
-    const shift = (fontSize * constants.axisHeight.value) / 1000;
-
-    // If useDisplayStyle is false then we need to reduce the font size of
-    // numerators and denominators
-    const useDisplayStyle = mathStyle === MathStyle.Display;
-
-    const minDenGap = useDisplayStyle
-        ? (fontSize * constants.fractionDenomDisplayStyleGapMin.value) / 1000
-        : (fontSize * constants.fractionDenominatorGapMin.value) / 1000;
-
-    const minNumGap = useDisplayStyle
-        ? (fontSize * constants.fractionNumDisplayStyleGapMin.value) / 1000
-        : (fontSize * constants.fractionNumeratorGapMin.value) / 1000;
-
-    // TODO: fix this code once we have debugging outlines in place, right now
-    // gap between the fraction bar and the numerator/denominator is too big
-    // when the font has been scaled.
-    // Check context.renderMode === RenderMode.Static when doing so.
-
-    // const numeratorShift = useDisplayStyle
-    //     ? (fontSize * constants.fractionNumeratorDisplayStyleShiftUp) / 1000
-    //     : (fontSize * constants.fractionNumeratorShiftUp) / 1000;
-
-    // const denominatorShift = useDisplayStyle
-    //     ? (fontSize * constants.fractionDenominatorDisplayStyleShiftDown) / 1000
-    //     : (fontSize * constants.fractionDenominatorShiftDown) / 1000;
-
-    // const numGap = Math.max(numeratorShift - numBox.depth - shift, minNumGap);
-    // const denGap = Math.max(
-    //     shift + denominatorShift - denBox.height,
-    //     minDenGap,
-    // );
-
-    const multiplier = multiplierForContext(context);
-    const endPadding = thickness; // add extra space around the numerator and denominator
-    const width =
-        Math.max(
-            Math.max(getWidth(numBox), getWidth(denBox)), // TODO: calculate this based on current font size
-            30 * multiplier, // empty numerator/denominator width
-        ) +
-        2 * endPadding;
-
-    // center the numerator
-    if (getWidth(numBox) < width) {
-        const kernSize = (width - getWidth(numBox)) / 2;
-        numBox = rebox(
-            numBox,
-            makeKern(kernSize, "start"),
-            makeKern(kernSize, "end"),
-        );
-    }
-
-    // center the denominator
-    if (getWidth(denBox) < width) {
-        const kernSize = (width - getWidth(denBox)) / 2;
-        denBox = rebox(
-            denBox,
-            makeKern(kernSize, "start"),
-            makeKern(kernSize, "end"),
-        );
-    }
-
-    const upList = makeList(minNumGap, numBox);
-    const dnList = makeList(minDenGap, denBox);
-    const stroke = hpackNat([[makeHRule(thickness, width)]], context);
-
-    const fracBox = makeVBox(width, stroke, upList, dnList, context);
-    fracBox.shift = -shift;
-
-    return fracBox;
-};
-
-const getConstantValue = (
+export const getConstantValue = (
     constant: MathValueRecord,
     context: Context,
 ): number => {
@@ -409,158 +323,4 @@ const getConstantValue = (
     const fontSize = fontSizeForContext(context);
 
     return (constant.value * fontSize) / font.head.unitsPerEm;
-};
-
-export const makeSubSup = (
-    subBox: Box | null,
-    supBox: Box | null,
-    context: Context,
-    prevEditNode?: Editor.types.Node | Editor.Focus,
-    prevLayoutNode?: Node,
-): Box => {
-    if (!supBox && !subBox) {
-        throw new Error("at least one of supBox and subBox must be defined");
-    }
-
-    const {font} = context.fontData;
-
-    const width = Math.max(
-        supBox ? getWidth(supBox) : 0,
-        subBox ? getWidth(subBox) : 0,
-    );
-
-    // Some atoms gets wrapped in a box to add padding to them so we need to
-    // filter them out.  Anything else that's in a box is some sort of compound
-    // layout structure (frac, delimited, etc.) and should have its subscript
-    // and/or superscript positioned based on the size of the box.
-    if (prevEditNode?.type !== "atom" && prevLayoutNode?.type === "Box") {
-        const {
-            superscriptBaselineDropMax,
-            subscriptBaselineDropMin,
-        } = font.math.constants;
-
-        const baselineDropMax = getConstantValue(
-            superscriptBaselineDropMax,
-            context,
-        );
-        const baselineDropMin = getConstantValue(
-            subscriptBaselineDropMin,
-            context,
-        );
-
-        const upList = [];
-        const dnList = [];
-
-        if (supBox) {
-            const shift = prevLayoutNode.height;
-            const kernShift = shift - baselineDropMax;
-
-            upList.push(makeKern(kernShift));
-            upList.push(supBox);
-        }
-
-        if (subBox) {
-            const shift = prevLayoutNode.depth;
-            const kernSize = shift - baselineDropMin;
-
-            dnList.push(makeKern(kernSize));
-            dnList.push(subBox);
-        }
-
-        const referenceNode = makeKern(0); // empty reference node
-        const subsupBox = makeVBox(
-            width,
-            referenceNode,
-            upList,
-            dnList,
-            context,
-        );
-
-        return subsupBox;
-    }
-
-    const {
-        subscriptTopMax,
-        superscriptBottomMin,
-        subscriptShiftDown,
-        superscriptShiftUp,
-        subSuperscriptGapMin,
-        superscriptBottomMaxWithSubscript,
-    } = font.math.constants;
-
-    const bottomMin = getConstantValue(superscriptBottomMin, context);
-    const shiftUp = getConstantValue(superscriptShiftUp, context);
-    const topMax = getConstantValue(subscriptTopMax, context);
-    const shiftDown = getConstantValue(subscriptShiftDown, context);
-
-    const upList = [];
-
-    if (supBox) {
-        // compute shift in baseline of superscript
-        const shift = Math.max(shiftUp, supBox.depth + bottomMin);
-
-        // -supBox.depth is to align the baseline of the superscript with the
-        // baseline of the base.
-        // TODO: replace the up/dn list that makeVBox uses with something else
-        // that doesn't require this correction
-        const kernShift = -supBox.depth + shift;
-
-        upList.push(makeKern(kernShift));
-        upList.push(supBox);
-    }
-
-    const dnList = [];
-
-    if (subBox) {
-        // compute shift in baseline of subscript
-        let shift = Math.max(shiftDown, subBox.height - topMax);
-
-        if (supBox) {
-            const supBoxShift = Math.max(shiftUp, supBox.depth + bottomMin);
-
-            const supBottom = supBoxShift - supBox.depth;
-            const subTop = subBox.height - shift;
-
-            const gap = supBottom - subTop;
-            const gapMin = getConstantValue(subSuperscriptGapMin, context);
-
-            if (gap < gapMin) {
-                if (upList[0].type === "Kern" && upList[1].type === "Box") {
-                    // shift the superscript up to increase the gap
-                    const correction = gapMin - gap;
-                    upList[0].size += correction;
-
-                    // We can't use upList[0].size in the calculation since it
-                    // includes the initial -supBox.depth to align baselines.
-                    const supBottom =
-                        supBoxShift + correction - upList[1].depth;
-                    const supBottomMax = getConstantValue(
-                        superscriptBottomMaxWithSubscript,
-                        context,
-                    );
-
-                    if (supBottom > supBottomMax) {
-                        // shift both down to maintain to gap
-                        const correcion = supBottom - supBottomMax;
-                        upList[0].size -= correcion;
-                        shift += correcion; // down is positive for dnList
-                    }
-                }
-            }
-        }
-
-        // We start with -subBox.height to align the subscript's baseline with
-        // the baseline of the base it's attached to
-        // TODO: replace the up/dn list that makeVBox uses with something else
-        // that doesn't require this correction
-        const kernSize = -subBox.height + shift;
-
-        dnList.push(makeKern(kernSize));
-        dnList.push(subBox);
-    }
-
-    const referenceNode = makeKern(0); // empty reference node
-    const subsupBox = makeVBox(width, referenceNode, upList, dnList, context);
-
-    return subsupBox;
 };
